rpcserver: adding validateUserBalance() in sendPayment()
This function validateUserBalance() will receive the extracted payment
intent and compare it to the total balance in the user's wallet. If the
total balance cannot satisfy the payment amount an error is returned
"not enough funds". This function is placed in sendPayment() after
extractPaymentIntent(), this so it can receive the extracted payment
intent and also not block, allowing the next payment to be processed.